requirements.txt
pandas
numpy
sqlalchemy
pymysql
jupyter
matplotlib
python-dotenv
